libraries/requirements.txt
opencv-python
numpy
keras-ocr
scikit-image
pdf2image
tensorflow
transformers
Pillow
torch
timm
planar